Fix production deployment to hardware Remove rick from container and service names Change network names to singular
Setup btrfs subvolume naming convention to match local-path storageclasses
Each btrfs subvolume is mount to agent nodes under /mnt with the same name as the storage class that uses it.

Rename operator kustomize directories that come default with k3s
Renamed with -custom suffix to allow skipping the default yaml files, but not the yaml files generated from this project.