Simplify v4l2rtspserver kustomize sub-directories
Setting up rtsp0 services for in cluster routing

Services that exist outside of the rick host (k3s-agent-1 in this case)
cannot route network requests to.
* This is because the service network default gateway exists on the rick host. Requests for services on host rpi4-0 (node k3s-agent-1) fail.
* Due to the default gateway on host rick being unaware of IPs outside
  of the host
* Changes to address this are
    * Setting internalTrafficPolicy: Cluster
    * Type: LoadBalancer (Allows routing from the rick host, where
      loadbalancer services run)
    * Changing the service ports to differ from those exposed by the
      loadbalancer
The end result is that motion is now able to successfully connect to the
rtsp0 service via dns name.

The ingress routes for traefik have been updated, but still need work to
be routed. Currently traefik recieves the request but doesn't
successfully route the request.